@@ -15,9 +15,11 @@ export class EventApiService {
     page,
     take,
     day,
-    date,
+    startDate,
+    endDate,
     type,
-    team,
+    home,
+    away,
     calendar,
     sortField = '',
     sortOrder = 1,
@@ -27,9 +29,11 @@ export class EventApiService {
         page,
         take,
         ...(day ? { day } : null),
-        ...(date ? { date } : null),
+        ...(startDate ? { startDate } : null),
+        ...(endDate ? { endDate } : null),
         ...(type ? { type } : null),
-        ...(team ? { team } : null),
+        ...(home ? { home } : null),
+        ...(away ? { away } : null),
         ...(calendar ? { calendar } : null),
         ...(sortField ? { sortField, sortOrder } : null),
       },
