@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { HotelListComponent } from './hotel-list/hotel-list.component';
 import { HotelRoomsComponent } from './hotel-rooms/hotel-rooms.component';
-import { HotelMealsComponent } from './hotel-meals/hotel-meals.component';
+import { HotelTeamsComponent } from './hotel-teams/hotel-teams.component';
 
 const routes: Routes = [
   {
@@ -13,12 +13,12 @@ const routes: Routes = [
     path: ':id',
     children: [
       {
-        path: '',
+        path: 'rooms',
         component: HotelRoomsComponent,
       },
       {
-        path: 'meals',
-        component: HotelMealsComponent,
+        path: 'teams',
+        component: HotelTeamsComponent,
       },
     ],
   },
