@@ -1,19 +1,21 @@
 import { Component } from '@angular/core';
-import { FormGroup, FormControl, Validators } from '@angular/forms';
-import { DynamicDialogRef, DynamicDialogConfig } from 'primeng/dynamicdialog';
+import { FormControl, FormGroup, Validators } from '@angular/forms';
+import { DynamicDialogConfig, DynamicDialogRef } from 'primeng/dynamicdialog';
+import { IMeal } from 'src/app/@core/models/meal.model';
 
 @Component({
-  selector: 'app-hotel-meals-modal',
-  templateUrl: './hotel-meals-modal.component.html',
-  styleUrls: ['./hotel-meals-modal.component.scss'],
+  selector: 'app-meal-modal',
+  templateUrl: './meal-modal.component.html',
+  styleUrls: ['./meal-modal.component.scss'],
 })
-export class HotelMealsModalComponent {
-  meal: any;
+export class MealModalComponent {
+  meal: Partial<IMeal> = {};
   isEdit: boolean = false;
 
   mealForm: FormGroup = new FormGroup({
     name: new FormControl('', Validators.required),
-    custom: new FormControl(false),
+    price: new FormControl(null),
+    maxConfigActive: new FormControl(null),
   });
 
   constructor(
@@ -31,7 +33,7 @@ export class HotelMealsModalComponent {
 
   onSave() {
     if (this.mealForm.valid) {
-      const meal = { ...this.mealForm.value };
+      const meal = this.mealForm.value;
       this.ref.close(meal);
     }
   }
