@@ -1,6 +1,8 @@
-import { Component } from '@angular/core';
-import { FormControl, FormGroup } from '@angular/forms';
+import { Component, DestroyRef, inject } from '@angular/core';
+import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { DynamicDialogConfig, DynamicDialogRef } from 'primeng/dynamicdialog';
+import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
+import { tap } from 'rxjs';
 
 @Component({
   selector: 'app-modal-meal',
@@ -11,79 +13,145 @@ export class ModalMealComponent {
   mealForm: FormGroup = new FormGroup({
     startDate: new FormControl(null),
     startDateHour: new FormControl(null),
-    quantity: new FormControl(1),
+    quantity: new FormControl(null),
     id: new FormControl(null),
     name: new FormControl(null),
+    mealId: new FormControl(null, Validators.required),
+    configId: new FormControl(null),
+    configIds: new FormControl([]),
     description: new FormControl(null),
-    custom: new FormControl(false),
   });
   isEdit: boolean = false;
   index: number = 0;
+  maxPax: number = 0;
   mealsList: any = [];
   isCustom = false;
+  private destroyRef = inject(DestroyRef);
+  selectedMeal: any;
+  selectedConfig: any;
+  selectedConfigs: any = [];
 
-  mealControl = new FormControl();
+  get mealIdControl() {
+    return this.mealForm.get('mealId') as FormControl;
+  }
+  get configIdControl() {
+    return this.mealForm.get('configId') as FormControl;
+  }
+  get configIdsControl() {
+    return this.mealForm.get('configIds') as FormControl;
+  }
+  get isConfigIdsError() {
+    return (
+      this.mealForm.get('configIds')?.value?.length >
+      this.selectedMeal?.maxConfigActive
+    );
+  }
 
   constructor(
     public ref: DynamicDialogRef,
     public config: DynamicDialogConfig
   ) {
     if (this.config.data) {
-      const { mealForm, mealsList, index, isEdit } = this.config.data;
-      this.mealForm = mealForm;
-      console.log('this.mealForm', this.mealForm);
+      const { mealForm, mealsList, index, isEdit, maxPax } = this.config.data;
+      this.mealForm.patchValue({ ...mealForm.value });
       this.mealsList = mealsList;
-      console.log('this.mealsList', this.mealsList);
       this.index = index;
+      this.maxPax = maxPax;
       this.isEdit = isEdit;
 
       if (this.isEdit) {
         const meal = this.mealForm.value;
-        console.log('meal', meal);
         const startDateHour = new Date(meal.startDate);
+
         startDateHour.setHours(
           new Date(meal.startDate).getHours(),
           new Date(meal.startDate).getMinutes()
         );
 
-        this.mealForm.addControl(
-          'startDateHour',
-          new FormControl(startDateHour)
-        );
+        this.mealForm.get('startDateHour')?.setValue(startDateHour);
 
-        this.mealControl.setValue({
-          id: meal.id,
-          name: meal.name,
-          custom: meal.custom,
-        });
+        if (this.mealIdControl.value) {
+          this.selectedMeal = this.mealsList.find(
+            (ml: any) => ml.id === this.mealIdControl.value
+          );
+        }
 
-        this.isCustom = meal.custom;
+        if (this.configIdControl.value) {
+          this.selectedConfig = this.selectedMeal.mealConfigs.find(
+            (mc: any) => mc.id === this.configIdControl.value
+          );
+        }
       }
     }
+
+    this.mealIdControl.valueChanges
+      .pipe(
+        takeUntilDestroyed(this.destroyRef),
+        tap((v) => {
+          this.selectedConfig = null;
+          this.configIdControl.reset();
+          this.configIdsControl.reset();
+          if (v) {
+            this.selectedMeal = this.mealsList.find((ml: any) => ml.id === v);
+          } else {
+            this.selectedMeal = null;
+          }
+        })
+      )
+      .subscribe();
+
+    this.configIdControl.valueChanges
+      .pipe(
+        takeUntilDestroyed(this.destroyRef),
+        tap((v) => {
+          if (v && this.selectedMeal) {
+            this.selectedConfig = this.selectedMeal.mealConfigs.find(
+              (mc: any) => mc.id === v
+            );
+          } else {
+            this.selectedConfig = null;
+            this.selectedConfigs = [];
+          }
+        })
+      )
+      .subscribe();
+
+    this.configIdsControl.valueChanges
+      .pipe(
+        takeUntilDestroyed(this.destroyRef),
+        tap((v) => {
+          if (v && this.selectedMeal) {
+            this.selectedConfigs = [...v].map((idConfig) =>
+              this.selectedMeal.mealConfigs.find(
+                (mc: any) => mc.id === idConfig
+              )
+            );
+          } else {
+            this.selectedConfigs = [];
+          }
+        })
+      )
+      .subscribe();
   }
 
   onSave() {
-    const { name, id } = this.mealControl.value;
-
+    // const { name, id } = this.mealControl.value;
     const meal = this.mealForm.value;
     const startDateHour = new Date(meal.startDateHour).getHours();
     const startDateMinute = new Date(meal.startDateHour).getMinutes();
     const startDate = new Date(meal.startDate);
     startDate.setHours(startDateHour, startDateMinute);
+
+    this.mealForm.removeControl('startDateHour');
     this.mealForm.patchValue({
-      ...meal,
       startDate,
-      name,
-      id,
-      custom: this.isCustom,
+      name: `${this.selectedMeal.name} - ${
+        this.selectedMeal.maxConfigActive === 1
+          ? this.selectedConfig.name
+          : this.selectedConfigs.map((sc: any) => sc.name).join(', ')
+      }`,
     });
 
-    console.log(this.mealForm.value);
     this.ref.close(this.mealForm);
-  }
-
-  onChangeMeal({ value }: any) {
-    console.log('value', value);
-    this.isCustom = value.custom;
   }
 }
