@@ -5,15 +5,18 @@ import { CartApiService } from 'src/app/@core/api/carts-api.service';
 import { MenuItem, Message, MessageService } from 'primeng/api';
 import { AuthService } from 'src/app/@core/services/auth.service';
 import { EventApiService } from 'src/app/@core/api/events-api.service';
-import { Subject, iif, map, switchMap, takeUntil, tap } from 'rxjs';
+import { Subject, forkJoin, iif, map, switchMap, takeUntil, tap } from 'rxjs';
 import { clearFormArray, uuidv4 } from 'src/app/@core/utils';
 import { DialogService, DynamicDialogRef } from 'primeng/dynamicdialog';
-import { CartCreateAccomodationsService } from './cart-create-accomodations/cart-create-accomodations.service';
-import { ECategoryPeople } from 'src/app/@core/models/people.model';
-import { EStatusCart } from 'src/app/@core/models/cart.model';
+import { PeopleRoomingService } from './people-rooming.service';
+import { ECategoryPeople, IPeople } from 'src/app/@core/models/people.model';
+import { EStatusCart, ICart } from 'src/app/@core/models/cart.model';
 import { HotelApiService } from 'src/app/@core/api/hotel-api.service';
 import { VeichleApiService } from 'src/app/@core/api/veichle-api.service';
 import { ActivityApiService } from 'src/app/@core/api/activity-api.service';
+import { IEvent } from 'src/app/@core/models/event.model';
+import { RoadApiService } from 'src/app/@core/api/road-api.service';
+import { MealApiService } from 'src/app/@core/api/meal-api.service';
 
 @Component({
   selector: 'app-personal-cart-create',
@@ -46,16 +49,20 @@ export class PersonalCartCreateComponent implements OnInit, OnDestroy {
   ];
   activeIndex: number = 0;
   currentUser: any;
-  event: any;
+  event!: Partial<IEvent>;
   city: any;
   hotels: any;
   veichles: any;
   hotelMeals: any;
-  cityActivities: any;
+  activities: any;
+  roads: any;
+  meals: any;
   EStatusCart = EStatusCart;
   status: EStatusCart = EStatusCart.DRAFT;
   messages: Message[] | undefined;
   selectedHotel: any = null;
+  people: IPeople[] = [];
+  cart!: Partial<ICart>;
 
   cartForm: FormGroup = new FormGroup({
     team: new FormGroup({
@@ -90,16 +97,10 @@ export class PersonalCartCreateComponent implements OnInit, OnDestroy {
     rooms: new FormArray([]),
     roads: new FormArray([]),
     meals: new FormArray([]),
+    people: new FormArray([]),
     genericNotes: new FormControl(null),
     accomodationNotes: new FormControl(null),
     roadNotes: new FormControl(null),
-  });
-  paxForm: FormGroup = new FormGroup({
-    players: new FormControl(null),
-    staffs: new FormControl(null),
-    managers: new FormControl(null),
-    equipments: new FormControl(null),
-    others: new FormControl(null),
   });
   accomodationForm: FormGroup = new FormGroup({
     hotel: new FormGroup({
@@ -133,20 +134,8 @@ export class PersonalCartCreateComponent implements OnInit, OnDestroy {
 
   ref!: DynamicDialogRef;
 
-  get activities() {
-    return this.cartForm.get('activities') as FormArray;
-  }
-
   get rooms() {
     return this.cartForm.get('rooms') as FormArray;
-  }
-
-  get roads() {
-    return this.cartForm.get('roads') as FormArray;
-  }
-
-  get meals() {
-    return this.cartForm.get('meals') as FormArray;
   }
 
   constructor(
@@ -156,11 +145,13 @@ export class PersonalCartCreateComponent implements OnInit, OnDestroy {
     private eventApiService: EventApiService,
     private hotelApiService: HotelApiService,
     private veichleApiService: VeichleApiService,
+    private roadApiService: RoadApiService,
     private activityApiService: ActivityApiService,
+    private mealApiService: MealApiService,
     private authService: AuthService,
     private messageService: MessageService,
     public dialogService: DialogService,
-    public accomodationService: CartCreateAccomodationsService
+    public peopleRoomingService: PeopleRoomingService
   ) {
     this.isEdit = route.snapshot.data['isEdit'];
     this.currentUser = this.authService.currentUser;
@@ -171,61 +162,99 @@ export class PersonalCartCreateComponent implements OnInit, OnDestroy {
         .pipe(
           takeUntil(this.unsubscribe$),
           switchMap(() => this.cartApiService.findOne(this.cartId)),
-          tap((cart) => this.patchFormEdit(cart)),
           tap((cart) => {
-            this.eventId = cart.event!.id!;
+            this.cart = { ...cart };
+            this.patchFormEdit();
+            this.event = { ...cart.event };
             this.status = cart.status!;
-          }),
-          switchMap(() => this.eventApiService.findOne(this.eventId)),
-          tap((event) => {
-            this.event = { ...event };
-            this.city = { ...this.event.home.city };
-            let roomings: any = [];
-            for (const room of (this.accomodationForm.get('rooms') as FormArray)
-              .value) {
-              for (const rooming of room.rooming) {
-                if (rooming.people && rooming.people.id) {
-                  roomings = [...roomings, rooming.people.id];
-                }
-              }
-            }
-            this.accomodationService.initPeople(
-              this.event.away.people ?? [],
-              roomings
-            );
+            this.city = { ...this.event.home?.city };
           }),
           switchMap(() =>
-            this.hotelApiService.findAll({
-              take: 200,
-              page: 1,
-              city: this.event.home.city.id,
-            })
-          ),
-          map((data) => data.data),
-          tap((hotels) => {
-            this.hotels = [...hotels];
-            this.selectedHotel = this.hotels.find(
-              (h: any) =>
-                h.id === this.accomodationForm.get('hotel')?.get('id')?.value
-            );
-            this.hotelMeals = [...(this.selectedHotel.meals ?? [])];
-            console.log('this hotel meals', this.hotelMeals);
-          }),
-          switchMap(() =>
-            this.veichleApiService.findAllRoad({
-              city: this.city.id,
-            })
-          ),
-          tap((veichles: any) => (this.veichles = [...veichles])),
-          switchMap(() =>
-            this.activityApiService.findAll({
-              take: 500,
-              page: 1,
-              city: this.city.id,
-            })
-          ),
-          map((data) => data.data),
-          tap((activities) => (this.cityActivities = [...activities]))
+            forkJoin({
+              hotels: this.hotelApiService
+                .findAll({
+                  take: 500,
+                  page: 1,
+                  teams: [this.event.home?.id],
+                })
+                .pipe(map((data) => data.data)),
+              veichles: this.veichleApiService
+                .findAll({ take: 500, page: 1 })
+                .pipe(map((data) => data.data)),
+              roads: this.roadApiService
+                .findAll({
+                  take: 500,
+                  page: 1,
+                  teams: [this.event.home?.id],
+                })
+                .pipe(map((data) => data.data)),
+              activities: this.activityApiService
+                .findAll({
+                  take: 500,
+                  page: 1,
+                  teams: [this.event.home?.id],
+                })
+                .pipe(map((data) => data.data)),
+              meals: this.mealApiService
+                .findAll({
+                  take: 500,
+                  page: 1,
+                })
+                .pipe(map((data) => data.data)),
+            }).pipe(
+              tap(({ hotels, veichles, roads, activities, meals }) => {
+                this.hotels = [...hotels];
+                this.veichles = [...veichles];
+                this.roads = [...roads];
+                this.activities = [...activities];
+                this.meals = [...meals];
+
+                this.selectedHotel = this.hotels.find(
+                  (h: any) =>
+                    h.id ===
+                    this.accomodationForm.get('hotel')?.get('id')?.value
+                );
+              })
+            )
+          )
+          //switchMap(() => this.eventApiService.findOne(this.eventId)),
+          //tap((event) => {
+          //this.event = { ...event };
+          //this.city = { ...this.event.home?.city };
+
+          //})
+          // switchMap(() =>
+          //   this.hotelApiService.findAll({
+          //     take: 200,
+          //     page: 1,
+          //     city: this.event.home?.city.id,
+          //   })
+          // ),
+          // map((data) => data.data),
+          // tap((hotels) => {
+          //   this.hotels = [...hotels];
+          //   this.selectedHotel = this.hotels.find(
+          //     (h: any) =>
+          //       h.id === this.accomodationForm.get('hotel')?.get('id')?.value
+          //   );
+          //   this.hotelMeals = [...(this.selectedHotel.meals ?? [])];
+          //   console.log('this hotel meals', this.hotelMeals);
+          // }),
+          // switchMap(() =>
+          //   this.veichleApiService.findAll({
+          //     city: this.city.id,
+          //   })
+          // ),
+          // tap((veichles: any) => (this.veichles = [...veichles])),
+          // switchMap(() =>
+          //   this.activityApiService.findAll({
+          //     take: 500,
+          //     page: 1,
+          //     city: this.city.id,
+          //   })
+          // ),
+          // map((data) => data.data),
+          // tap((activities) => (this.cityActivities = [...activities]))
         )
         .subscribe();
     } else {
@@ -236,37 +265,55 @@ export class PersonalCartCreateComponent implements OnInit, OnDestroy {
           switchMap(() => this.eventApiService.findOne(this.eventId)),
           tap((event) => {
             this.event = { ...event };
-            this.city = { ...this.event.home.city };
-            this.accomodationService.initPeople(
-              this.event.away.people ?? [],
+            this.city = { ...this.event.home?.city };
+            this.people = [...(this.event.away?.people ?? [])];
+            this.peopleRoomingService.initPeople(
+              this.event.away?.people ?? [],
               []
             );
-            this.setPax();
           }),
           switchMap(() =>
-            this.hotelApiService.findAll({
-              take: 200,
-              page: 1,
-              city: this.event.home.city.id,
-            })
-          ),
-          map((data) => data.data),
-          tap((hotels) => (this.hotels = [...hotels])),
-          switchMap(() =>
-            this.veichleApiService.findAllRoad({
-              city: this.city.id,
-            })
-          ),
-          tap((veichles: any) => (this.veichles = [...veichles])),
-          switchMap(() =>
-            this.activityApiService.findAll({
-              take: 500,
-              page: 1,
-              city: this.city.id,
-            })
-          ),
-          map((data) => data.data),
-          tap((activities) => (this.cityActivities = [...activities]))
+            forkJoin({
+              hotels: this.hotelApiService
+                .findAll({
+                  take: 500,
+                  page: 1,
+                  teams: [this.event.home?.id],
+                })
+                .pipe(map((data) => data.data)),
+              veichles: this.veichleApiService
+                .findAll({ take: 500, page: 1 })
+                .pipe(map((data) => data.data)),
+              roads: this.roadApiService
+                .findAll({
+                  take: 500,
+                  page: 1,
+                  teams: [this.event.home?.id],
+                })
+                .pipe(map((data) => data.data)),
+              activities: this.activityApiService
+                .findAll({
+                  take: 500,
+                  page: 1,
+                  teams: [this.event.home?.id],
+                })
+                .pipe(map((data) => data.data)),
+              meals: this.mealApiService
+                .findAll({
+                  take: 500,
+                  page: 1,
+                })
+                .pipe(map((data) => data.data)),
+            }).pipe(
+              tap(({ hotels, veichles, roads, activities, meals }) => {
+                this.hotels = [...hotels];
+                this.veichles = [...veichles];
+                this.roads = [...roads];
+                this.activities = [...activities];
+                this.meals = [...meals];
+              })
+            )
+          )
         )
         .subscribe();
     }
@@ -343,22 +390,36 @@ export class PersonalCartCreateComponent implements OnInit, OnDestroy {
     }
   }
 
-  setPax() {
-    const keyBinding = [
-      { key: 'players', category: ECategoryPeople.PLAYER },
-      { key: 'staffs', category: ECategoryPeople.STAFF },
-      { key: 'managers', category: ECategoryPeople.MANAGER },
-      { key: 'equipments', category: ECategoryPeople.EQUIPMENT },
-      { key: 'others', category: ECategoryPeople.OTHER },
-    ];
+  onAddNewPax(pax: IPeople) {
+    const peopleCopy = [...this.people, pax];
+    this.people = [...peopleCopy];
+    this.peopleRoomingService.updatePeople(pax, 'ADD');
+  }
 
-    for (const k of keyBinding) {
-      this.paxForm
-        .get(k.key)
-        ?.setValue(
-          this.event.away.people.filter((p: any) => p.category === k.category)
-            .length ?? 0
-        );
+  onRemovePax(pax: IPeople) {
+    const peopleCopy = this.people.filter(
+      (p) => !this.peopleRoomingService.checkPaxAreEquals(p, pax)
+    );
+    this.people = [...peopleCopy];
+    this.peopleRoomingService.updatePeople(pax, 'REMOVE');
+
+    const rooms = (this.accomodationForm.get('rooms') as FormArray).value;
+    const indexRoom = rooms.findIndex((room: any) =>
+      room.rooming.some((r: any) =>
+        this.peopleRoomingService.checkPaxAreEquals(r, pax)
+      )
+    );
+
+    if (indexRoom > -1) {
+      const indexRooming = rooms[indexRoom].rooming.findIndex((r: any) =>
+        this.peopleRoomingService.checkPaxAreEquals(r, pax)
+      );
+
+      (
+        (this.accomodationForm.get('rooms') as FormArray)
+          .at(indexRoom)
+          .get('rooming') as FormArray
+      ).removeAt(indexRooming);
     }
   }
 
@@ -366,30 +427,67 @@ export class PersonalCartCreateComponent implements OnInit, OnDestroy {
     clearFormArray(this.mealForm.get('meals') as FormArray);
     clearFormArray(this.accomodationForm.get('rooms') as FormArray);
 
-    this.hotelMeals = [...hotel.meals];
+    this.selectedHotel = { ...hotel };
+  }
+
+  onCopyRooming(data: any) {
+    console.log('onCopyRooming', data);
+    clearFormArray(this.rooms);
+
+    let setHotel = false;
+
+    for (const d of data.rooming) {
+      if (!setHotel) {
+        setHotel = true;
+        const hotel = this.hotels.find((h: any) => h.id === d.hotelId);
+        this.selectedHotel = { ...hotel };
+      }
+
+      const room = this.selectedHotel.rooms.find(
+        (r: any) => r.name === d.roomName
+      );
+      const newRoom = new FormGroup({
+        id: new FormControl(room.id),
+        hotelId: new FormControl(this.selectedHotel.id),
+        hotelName: new FormControl(this.selectedHotel.name),
+        name: new FormControl(room.name),
+        price: new FormControl(room.price),
+        numPax: new FormControl(room.numPax),
+        rooming: new FormArray([]),
+        uuid: new FormControl(uuidv4()),
+      });
+      for (let p of d.people) {
+        const newPeople = new FormGroup({
+          name: new FormControl(p.name),
+          surname: new FormControl(p.surname),
+          category: new FormControl(p.category),
+        });
+        (newRoom.get('rooming') as FormArray).push(newPeople);
+      }
+      (this.accomodationForm.get('rooms') as FormArray).push(newRoom);
+    }
+
+    this.people = [...data.people];
+
+    let roomings: any = [];
+    for (const room of (this.accomodationForm.get('rooms') as FormArray)
+      .value) {
+      for (const rooming of room.rooming) {
+        roomings = [...roomings, rooming];
+      }
+    }
+    this.peopleRoomingService.initPeople(this.people ?? [], roomings);
+
+    this.messageService.add({
+      severity: 'success',
+      summary: 'Rooming copiata da ultima trasferta',
+    });
   }
 
   onSaveCart() {
     const cart = { ...this.cartForm.value };
-
-    // this.ref = this.dialogService.open(PersonalCartCrateConfirmModalComponent, {
-    //   header: this.isEdit ? 'Aggiornamento trasferta' : 'Creazione trasferta',
-    //   width: '450px',
-    //   contentStyle: { overflow: 'visible' },
-    //   baseZIndex: 10001,
-    // });
-
-    // this.ref.onClose.subscribe((action: string) => {
-    //   switch (action) {
-    //     case 'confirm':
-    //       cart.isCompleted = true;
-    //       break;
-    //     default:
-    //       break;
-    //   }
-
+    console.log('SAVE CART', cart);
     this.save$.next(cart);
-    //});
   }
 
   /**
@@ -397,54 +495,42 @@ export class PersonalCartCreateComponent implements OnInit, OnDestroy {
    * serve a popolare i form per le tab a partire dal
    * carrello recuperato dalle API
    */
-  patchFormEdit(cart: any) {
-    console.log('patchFormEdit', cart);
+  patchFormEdit() {
+    console.log('patchFormEdit', this.cart);
     // PATCH CART FORM
-    this.cartForm.get('genericNotes')?.setValue(cart.genericNotes);
+    this.cartForm.get('genericNotes')?.setValue(this.cart.genericNotes);
     // PATCH PAX FORM
-    const { players, managers, staffs, equipments, others } = cart;
-    this.paxForm.patchValue({ players, managers, staffs, equipments, others });
+    const { players, managers, staffs, equipments, others, people } = this.cart;
+    this.people = [...(people ?? [])];
 
     // PATCH ACCOMODATION FORM
-    const startDate = cart.startDate ? new Date(cart.startDate) : null;
-    const endDate = cart.endDate ? new Date(cart.endDate) : null;
+    const startDate = this.cart.startDate
+      ? new Date(this.cart.startDate)
+      : null;
+    const endDate = this.cart.endDate ? new Date(this.cart.endDate) : null;
     this.accomodationForm.patchValue({
       startDate,
       endDate,
-      accomodationNotes: cart.accomodationNotes,
+      accomodationNotes: this.cart.accomodationNotes,
     });
-    if (cart.rooms && cart.rooms.length) {
-      const { id, name } = cart.rooms[0].room.hotel;
+    if (this.cart.rooms && this.cart.rooms.length) {
+      const { id, name } = this.cart.rooms[0].room?.hotel;
       this.accomodationForm.patchValue({ hotel: { id, name } });
       const accomodationRooms = this.accomodationForm.get('rooms') as FormArray;
-      for (const r of cart.rooms) {
+      for (const r of this.cart.rooms) {
         const room = new FormGroup({
-          id: new FormControl(r.room.id),
-          name: new FormControl(r.room.name),
-          price: new FormControl(r.room.price),
-          numPax: new FormControl(r.room.numPax),
+          id: new FormControl(r.room?.id),
+          name: new FormControl(r.room?.name),
+          price: new FormControl(r.room?.price),
+          numPax: new FormControl(r.room?.numPax),
           rooming: new FormArray([]),
-          hotelId: new FormControl(r.room.hotel.id),
-          hotelName: new FormControl(r.room.hotel.name),
+          hotelId: new FormControl(r.room?.hotel.id),
+          hotelName: new FormControl(r.room?.hotel.name),
           uuid: new FormControl(uuidv4()),
         });
 
-        for (const cartRooming of r.rooming) {
+        for (const cartRooming of r!.rooming!) {
           const rooming = new FormGroup({
-            people: new FormGroup({
-              id: new FormControl(
-                cartRooming.people ? cartRooming.people.id : null
-              ),
-              name: new FormControl(
-                cartRooming.people ? cartRooming.people.name : null
-              ),
-              surname: new FormControl(
-                cartRooming.people ? cartRooming.people.surname : null
-              ),
-              category: new FormControl(
-                cartRooming.people ? cartRooming.people.category : null
-              ),
-            }),
             name: new FormControl(cartRooming.name),
             surname: new FormControl(cartRooming.surname),
             category: new FormControl(cartRooming.category),
@@ -455,18 +541,70 @@ export class PersonalCartCreateComponent implements OnInit, OnDestroy {
       }
     }
 
+    let roomings: any = [];
+    for (const room of (this.accomodationForm.get('rooms') as FormArray)
+      .value) {
+      for (const rooming of room.rooming) {
+        roomings = [...roomings, rooming];
+      }
+    }
+    this.peopleRoomingService.initPeople(this.cart.people ?? [], roomings);
+
     // PATCH MEAL FORM
-    if (cart.meals && cart.meals.length) {
+    if (this.cart.meals && this.cart.meals.length) {
+      const mealsOnCarts = this.cart.meals.reduce((group: any, meal: any) => {
+        const { quantity, startDate, description } = meal;
+        const { id: configId, name: configName } = meal.meal;
+
+        const { id: mealId, name: mealName } = meal.meal.meal;
+        const index = group.findIndex((g: any) => g.mealId === mealId);
+        if (index > -1) {
+          group[index].configIds.push({ configId, configName });
+        } else {
+          group.push({
+            mealId,
+            mealName,
+            configIds: [{ configId, configName }],
+            quantity,
+            startDate,
+            description,
+          });
+        }
+        return group;
+      }, []);
       const meals = this.mealForm.get('meals') as FormArray;
-      for (const m of cart.meals) {
+      for (const m of mealsOnCarts) {
         console.log('m', m);
         const room = new FormGroup({
-          id: new FormControl(m.meal.id),
-          name: new FormControl(m.meal.name),
+          id: new FormControl(m.mealId),
           quantity: new FormControl(m.quantity),
-          description: new FormControl(m.description),
           startDate: new FormControl(new Date(m.startDate)),
-          custom: new FormControl(m.meal.custom),
+          mealId: new FormControl(m.mealId),
+          ...(m.configIds.length < 2
+            ? {
+                configId: new FormControl(m.configIds[0].configId),
+                configIds: new FormControl([]),
+              }
+            : {
+                configId: new FormControl(null),
+                configIds: new FormControl(
+                  m.configIds.map((c: any) => c.configId)
+                ),
+              }),
+          ...(m.configIds.length < 2
+            ? {
+                name: new FormControl(
+                  `${m.mealName} - ${m.configIds[0].configName}`
+                ),
+              }
+            : {
+                name: new FormControl(
+                  `${m.mealName} - ${m.configIds
+                    .map((c: any) => c.configName)
+                    .join(', ')}`
+                ),
+              }),
+          description: new FormControl(null),
         });
         meals.push(room);
       }
@@ -474,23 +612,27 @@ export class PersonalCartCreateComponent implements OnInit, OnDestroy {
 
     // PATCH ROADS FORM
     this.roadForm.patchValue({
-      roadNotes: cart.roadNotes,
+      roadNotes: this.cart.roadNotes,
     });
-    if (cart.roads && cart.roads.length) {
+    if (this.cart.roads && this.cart.roads.length) {
       const roads = this.roadForm.get('roads') as FormArray;
-      for (const r of cart.roads) {
+      for (const r of this.cart.roads) {
         const road = new FormGroup({
-          id: new FormControl(r.road.id),
-          startDate: new FormControl(new Date(r.startDate)),
-          startDateHour: new FormControl(new Date(r.startDate)),
-          price: new FormControl(r.road.price),
-          from: new FormControl(r.road.from),
-          to: new FormControl(r.road.to),
+          startDate: new FormControl(new Date(r.startDate!)),
+          startDateHour: new FormControl(new Date(r.startDate!)),
           quantity: new FormControl(r.quantity),
-          veichleId: new FormControl(r.road.veichle.id),
-          veichle: new FormControl(r.road.veichle.name),
-          createdAt: new FormControl(r.road.createdAt),
-          updatedAt: new FormControl(r.road.updatedAt),
+          road: new FormGroup({
+            price: new FormControl(r.road?.price),
+            from: new FormControl(r.road?.from),
+            to: new FormControl(r.road?.to),
+            id: new FormControl(r.road?.id),
+          }),
+          veichle: new FormGroup({
+            id: new FormControl(r.veichle?.id),
+            name: new FormControl(r.veichle?.name),
+          }),
+          createdAt: new FormControl(r.createdAt!),
+          updatedAt: new FormControl(r.updatedAt!),
         });
 
         roads.push(road);
@@ -498,24 +640,28 @@ export class PersonalCartCreateComponent implements OnInit, OnDestroy {
     }
 
     // PATCH ACTIVITY FORM
-    if (cart.activities && cart.activities.length) {
+    if (this.cart.activities && this.cart.activities.length) {
       const actvitiyActivities = this.activityForm.get(
         'activities'
       ) as FormArray;
-      for (const a of cart.activities) {
+      for (const a of this.cart.activities) {
         const activity = new FormGroup({
-          id: new FormControl(a.activity.id),
-          name: new FormControl(a.activity.name),
-          description: new FormControl(a.activity.description),
-          price: new FormControl(a.activity.price),
+          id: new FormControl(a.activity?.id),
+          name: new FormControl(a.activity?.name),
+          description: new FormControl(a.activity?.description),
+          price: new FormControl(a.activity?.price),
           quantity: new FormControl(a.quantity),
-          startDate: new FormControl(new Date(a.startDate)),
+          startDate: new FormControl(new Date(a.startDate!)),
         });
         actvitiyActivities.push(activity);
       }
     }
 
-    console.log('after patch cart', this.cartForm.value);
+    console.log('AFTER PATCH CART');
+    console.log('ACCOMODATION', this.accomodationForm.value);
+    console.log('MEAL', this.mealForm.value);
+    console.log('ACTIVITY', this.activityForm.value);
+    console.log('ROAD', this.roadForm.value);
   }
 
   /**
@@ -524,8 +670,23 @@ export class PersonalCartCreateComponent implements OnInit, OnDestroy {
    * il form principale del carrello
    */
   populateCartForm() {
-    console.log('populate cart form');
-    this.cartForm.patchValue(this.paxForm.value);
+    this.cartForm.patchValue({
+      players: (
+        this.people.filter((p) => p.category == ECategoryPeople.PLAYER) ?? []
+      ).length,
+      staffs: (
+        this.people.filter((p) => p.category == ECategoryPeople.STAFF) ?? []
+      ).length,
+      managers: (
+        this.people.filter((p) => p.category == ECategoryPeople.MANAGER) ?? []
+      ).length,
+      equipments: (
+        this.people.filter((p) => p.category == ECategoryPeople.EQUIPMENT) ?? []
+      ).length,
+      others: (
+        this.people.filter((p) => p.category == ECategoryPeople.OTHER) ?? []
+      ).length,
+    });
 
     this.cartForm.get('event')?.get('id')?.setValue(this.event.id);
     this.cartForm.get('event')?.get('date')?.setValue(this.event.date);
@@ -533,38 +694,38 @@ export class PersonalCartCreateComponent implements OnInit, OnDestroy {
       .get('event')
       ?.get('home')
       ?.get('id')
-      ?.setValue(this.event.home.id);
+      ?.setValue(this.event.home?.id);
     this.cartForm
       .get('event')
       ?.get('home')
       ?.get('name')
-      ?.setValue(this.event.home.name);
+      ?.setValue(this.event.home?.name);
     this.cartForm
       .get('event')
       ?.get('away')
       ?.get('id')
-      ?.setValue(this.event.away.id);
+      ?.setValue(this.event.away?.id);
     this.cartForm
       .get('event')
       ?.get('away')
       ?.get('name')
-      ?.setValue(this.event.away.name);
+      ?.setValue(this.event.away?.name);
 
-    this.cartForm.get('team')?.get('id')?.setValue(this.event.away.id);
-    this.cartForm.get('team')?.get('name')?.setValue(this.event.away.name);
+    this.cartForm.get('team')?.get('id')?.setValue(this.event.away?.id);
+    this.cartForm.get('team')?.get('name')?.setValue(this.event.away?.name);
 
     this.cartForm
       .get('event')
       ?.get('home')
       ?.get('city')
       ?.get('id')
-      ?.setValue(this.event.home.city.id);
+      ?.setValue(this.event.home?.city.id);
     this.cartForm
       .get('event')
       ?.get('home')
       ?.get('city')
       ?.get('name')
-      ?.setValue(this.event.home.city.name);
+      ?.setValue(this.event.home?.city.name);
 
     const { startDate, endDate, accomodationNotes } =
       this.accomodationForm.value;
@@ -576,6 +737,23 @@ export class PersonalCartCreateComponent implements OnInit, OnDestroy {
     const roadRoads = this.roadForm.get('roads') as FormArray;
     const activiyActivities = this.activityForm.get('activities') as FormArray;
     const mealMeals = this.mealForm.get('meals') as FormArray;
+
+    // POPULATE PEOPLE
+    clearFormArray(this.cartForm.get('people') as FormArray);
+    for (const p of this.people) {
+      const cartPeople = this.cartForm.get('people') as FormArray;
+      cartPeople.push(
+        new FormGroup({
+          birthDate: new FormControl(p.birthDate),
+          birthPlace: new FormControl(p.birthPlace),
+          category: new FormControl(p.category),
+          docExpiredAt: new FormControl(p.docExpiredAt),
+          docNumber: new FormControl(p.docNumber),
+          name: new FormControl(p.name),
+          surname: new FormControl(p.surname),
+        })
+      );
+    }
 
     // PUPOLATE ROOMS
     clearFormArray(this.cartForm.get('rooms') as FormArray);
@@ -604,16 +782,18 @@ export class PersonalCartCreateComponent implements OnInit, OnDestroy {
     this.cartForm.get('roadNotes')?.setValue(roadNotes);
     clearFormArray(this.cartForm.get('roads') as FormArray);
     for (const r of roadRoads.value) {
+      console.log('R', r);
       const cartRoads = this.cartForm.get('roads') as FormArray;
       cartRoads.push(
         new FormGroup({
-          roadId: new FormControl(r.id),
+          roadId: new FormControl(r.road.id),
           quantity: new FormControl(r.quantity),
           startDate: new FormControl(r.startDate),
-          veichleName: new FormControl(r.veichle),
-          from: new FormControl(r.from),
-          to: new FormControl(r.to),
-          price: new FormControl(r.price),
+          veichleId: new FormControl(r.veichle.id),
+          veichle: new FormControl(r.veichle.name),
+          from: new FormControl(r.road.from),
+          to: new FormControl(r.road.to),
+          price: new FormControl(r.road.price),
         })
       );
     }
@@ -622,7 +802,6 @@ export class PersonalCartCreateComponent implements OnInit, OnDestroy {
     clearFormArray(this.cartForm.get('activities') as FormArray);
     for (const a of activiyActivities.value) {
       const cartActivities = this.cartForm.get('activities') as FormArray;
-
       cartActivities.push(
         new FormGroup({
           activityId: new FormControl(a.id),
@@ -639,15 +818,31 @@ export class PersonalCartCreateComponent implements OnInit, OnDestroy {
     clearFormArray(this.cartForm.get('meals') as FormArray);
     for (const m of mealMeals.value) {
       const cartMeals = this.cartForm.get('meals') as FormArray;
-      cartMeals.push(
-        new FormGroup({
-          mealId: new FormControl(m.id),
-          quantity: new FormControl(m.quantity),
-          startDate: new FormControl(m.startDate),
-          name: new FormControl(m.name),
-          description: new FormControl(m.description),
-        })
-      );
+      if (m.configIds && m.configIds.length) {
+        for (const config of m.configIds) {
+          cartMeals.push(
+            new FormGroup({
+              mealId: new FormControl(m.mealId),
+              configId: new FormControl(config),
+              quantity: new FormControl(m.quantity),
+              startDate: new FormControl(m.startDate),
+              name: new FormControl(m.name),
+              description: new FormControl(m.description),
+            })
+          );
+        }
+      } else {
+        cartMeals.push(
+          new FormGroup({
+            mealId: new FormControl(m.mealId),
+            configId: new FormControl(m.configId),
+            quantity: new FormControl(m.quantity),
+            startDate: new FormControl(m.startDate),
+            name: new FormControl(m.name),
+            description: new FormControl(m.description),
+          })
+        );
+      }
     }
 
     console.log('CREAZIONE CART', this.cartForm.value);
