@@ -12,6 +12,8 @@ import { VeichleModalComponent } from '../veichle-modal/veichle-modal.component'
 import { CityApiService } from 'src/app/@core/api/city-api.service';
 import { OverlayPanel } from 'primeng/overlaypanel';
 import { Router } from '@angular/router';
+import { TableLazyLoadEvent } from 'primeng/table';
+import { ISort } from 'src/app/@core/models/base.model';
 
 @Component({
   selector: 'app-veichle-list',
@@ -25,7 +27,7 @@ export class VeichleListComponent {
   page: number = 0;
   size: number = 10;
   filter: string = '';
-  sort: any = null;
+  sort: ISort | null = null;
   loading: boolean = false;
   filterActive: boolean = false;
 
@@ -117,13 +119,13 @@ export class VeichleListComponent {
     this.veichles$.next();
   }
 
-  onChangePage(event: LazyLoadEvent) {
+  onChangePage(event: TableLazyLoadEvent) {
     this.page = event.first! / event.rows! || 0;
 
     if (event.sortField) {
       this.sort = {
         field: event.sortField,
-        order: event.sortOrder,
+        order: event.sortOrder ?? 0,
       };
     } else {
       this.sort = null;
